fix(recommendations): prepping recommendations to fire impression events for analytics
